Chore (./contracts): Linging && Fixing wring bytecode

+++ contracts/src/verifiers/PartialWithdrawVerifier.sol
@@ -22,17 +22,17 @@ pragma solidity >=0.7.0 <0.9.0;
 
 contract Groth16Verifier {
     // Scalar field size
-    uint256 constant r    = 21888242871839275222246405745257275088548364400416034343698204186575808495617;
+    uint256 constant r = 21888242871839275222246405745257275088548364400416034343698204186575808495617;
     // Base field size
-    uint256 constant q   = 21888242871839275222246405745257275088696311157297823662689037894645226208583;
+    uint256 constant q = 21888242871839275222246405745257275088696311157297823662689037894645226208583;
 
     // Verification Key data
-    uint256 constant alphax  = 7551512397760434695278787407970089361920976085724591893631715113665513493155;
-    uint256 constant alphay  = 4858743534958594702299513812002532330030483289133666508947444335896973309544;
-    uint256 constant betax1  = 7138136986218441300139150885539646496770234652739011163448551918795965653539;
-    uint256 constant betax2  = 4712825566973883170357105386159418790555493841853209642048622776393876364613;
-    uint256 constant betay1  = 19907689548832711461939838245098301347723498620225312171842325865135980934261;
-    uint256 constant betay2  = 6113291132796547763696096414016747469354975217717077088790776678006800547003;
+    uint256 constant alphax = 7551512397760434695278787407970089361920976085724591893631715113665513493155;
+    uint256 constant alphay = 4858743534958594702299513812002532330030483289133666508947444335896973309544;
+    uint256 constant betax1 = 7138136986218441300139150885539646496770234652739011163448551918795965653539;
+    uint256 constant betax2 = 4712825566973883170357105386159418790555493841853209642048622776393876364613;
+    uint256 constant betay1 = 19907689548832711461939838245098301347723498620225312171842325865135980934261;
+    uint256 constant betay2 = 6113291132796547763696096414016747469354975217717077088790776678006800547003;
     uint256 constant gammax1 = 11559732032986387107991004021392285783925812861821192530917403151452391805634;
     uint256 constant gammax2 = 10857046999023057135944570762232829481370756359578518086990519993285655852781;
     uint256 constant gammay1 = 4082367875863433681332203403145435568316851327593401208105741076214120093531;
@@ -42,42 +42,45 @@ contract Groth16Verifier {
     uint256 constant deltay1 = 4082367875863433681332203403145435568316851327593401208105741076214120093531;
     uint256 constant deltay2 = 8495653923123431417604973247489272438418190587263600148770280649306958101930;
 
-    
     uint256 constant IC0x = 9859531802762154984017284397834533482993524067642907941860845497134235603027;
     uint256 constant IC0y = 4165015840159009947285286895330207397446057080673444564843324975445955801498;
-    
+
     uint256 constant IC1x = 1652234167291409770770648243277049061805859192645730841155890203424899184644;
     uint256 constant IC1y = 15749642841203088461813725893453859939958977191427502134683785217961773810222;
-    
+
     uint256 constant IC2x = 10600794894627056142825634124498424658931845524569468319467344908527138244154;
     uint256 constant IC2y = 1178453042819229304775691900396718569623275366922439611410348810021956258996;
-    
+
     uint256 constant IC3x = 19862803396237877304323898782776574309993806333303064574415130316126090840853;
     uint256 constant IC3y = 7771196390432921437676178762452898668335706631953569848590136656681619339108;
-    
+
     uint256 constant IC4x = 20020581519657994844620070862534002194612041268114382416550830654606323959245;
     uint256 constant IC4y = 1582484238120615129692911636156033685802356984032139196848843815688812524494;
-    
+
     uint256 constant IC5x = 8254071638760208207495249019163695958771296252673902558605446018190980813242;
     uint256 constant IC5y = 9503162588005506839444096674766766910160760148413632677923332040360936415277;
-    
+
     uint256 constant IC6x = 8863891880680466341394831404729892450114987314096338397619109299058206345292;
     uint256 constant IC6y = 3136434610472477726925568912258597724277203657924287717733271681885673351728;
-    
+
     uint256 constant IC7x = 4462148082242382410740052693446252382591144769111058869034925653249196978211;
     uint256 constant IC7y = 1037593869680256238499383982835650202286022101321033773775535566392417918911;
-    
+
     uint256 constant IC8x = 16321234651301401282634066627773075550139281805625289561658032663225671547560;
     uint256 constant IC8y = 18307754989848079155698353131040277437904839172105944991022849763067492496492;
-    
- 
+
     // Memory data
     uint16 constant pVk = 0;
     uint16 constant pPairing = 128;
 
     uint16 constant pLastMem = 896;
 
-    function verifyProof(uint[2] calldata _pA, uint[2][2] calldata _pB, uint[2] calldata _pC, uint[8] calldata _pubSignals) public view returns (bool) {
+    function verifyProof(
+        uint256[2] calldata _pA,
+        uint256[2][2] calldata _pB,
+        uint256[2] calldata _pC,
+        uint256[8] calldata _pubSignals
+    ) public view returns (bool) {
         assembly {
             function checkField(v) {
                 if iszero(lt(v, q)) {
@@ -85,7 +88,7 @@ contract Groth16Verifier {
                     return(0, 0x20)
                 }
             }
-            
+
             // G1 function to multiply a G1 value(x,y) to value in an address
             function g1_mulAccC(pR, x, y, s) {
                 let success
@@ -120,23 +123,22 @@ contract Groth16Verifier {
                 mstore(add(_pVk, 32), IC0y)
 
                 // Compute the linear combination vk_x
-                
+
                 g1_mulAccC(_pVk, IC1x, IC1y, calldataload(add(pubSignals, 0)))
-                
+
                 g1_mulAccC(_pVk, IC2x, IC2y, calldataload(add(pubSignals, 32)))
-                
+
                 g1_mulAccC(_pVk, IC3x, IC3y, calldataload(add(pubSignals, 64)))
-                
+
                 g1_mulAccC(_pVk, IC4x, IC4y, calldataload(add(pubSignals, 96)))
-                
+
                 g1_mulAccC(_pVk, IC5x, IC5y, calldataload(add(pubSignals, 128)))
-                
+
                 g1_mulAccC(_pVk, IC6x, IC6y, calldataload(add(pubSignals, 160)))
-                
+
                 g1_mulAccC(_pVk, IC7x, IC7y, calldataload(add(pubSignals, 192)))
-                
+
                 g1_mulAccC(_pVk, IC8x, IC8y, calldataload(add(pubSignals, 224)))
-                
 
                 // -A
                 mstore(_pPairing, calldataload(pA))
@@ -162,7 +164,6 @@ contract Groth16Verifier {
                 mstore(add(_pPairing, 384), mload(add(pMem, pVk)))
                 mstore(add(_pPairing, 416), mload(add(pMem, add(pVk, 32))))
 
-
                 // gamma2
                 mstore(add(_pPairing, 448), gammax1)
                 mstore(add(_pPairing, 480), gammax2)
@@ -179,7 +180,6 @@ contract Groth16Verifier {
                 mstore(add(_pPairing, 704), deltay1)
                 mstore(add(_pPairing, 736), deltay2)
 
-
                 let success := staticcall(sub(gas(), 2000), 8, _pPairing, 768, _pPairing, 0x20)
 
                 isOk := and(success, mload(_pPairing))
@@ -189,31 +189,30 @@ contract Groth16Verifier {
             mstore(0x40, add(pMem, pLastMem))
 
             // Validate that all evaluations ∈ F
-            
+
             checkField(calldataload(add(_pubSignals, 0)))
-            
+
             checkField(calldataload(add(_pubSignals, 32)))
-            
+
             checkField(calldataload(add(_pubSignals, 64)))
-            
+
             checkField(calldataload(add(_pubSignals, 96)))
-            
+
             checkField(calldataload(add(_pubSignals, 128)))
-            
+
             checkField(calldataload(add(_pubSignals, 160)))
-            
+
             checkField(calldataload(add(_pubSignals, 192)))
-            
+
             checkField(calldataload(add(_pubSignals, 224)))
-            
+
             checkField(calldataload(add(_pubSignals, 256)))
-            
 
             // Validate all evaluations
             let isValid := checkPairing(_pA, _pB, _pC, _pubSignals, pMem)
 
             mstore(0, isValid)
-             return(0, 0x20)
-         }
-     }
- }
+            return(0, 0x20)
+        }
+    }
+}
